pyproject.toml
[project]
name = "discord-virtual-micro"
version = "0.2.0"
description = "Discord Virtual Microphone is a bot that leverages AI from Silero Models to read out loud any text you type in Discord. This bot can be particularly useful for individuals without access to a physical microphone."
authors = [{ name = "Vazgen", email = "norvardyanvazgen@gmail.com" }]
dependencies = [
    "soundcard>=0.4.2",
    "librosa>=0.10.1",
    "discord.py-self @ git+https://github.com/dolfies/discord.py-self.git@9c1047bc1f5ae6a739dd246212d049d601bccfc3#egg=discord-py-self",
    "num2words>=0.5.13",
    "omegaconf>=2.3.0",
    "transliterate>=1.10.2",
    "setuptools>=69.1.1",
    "torch>=2.2.1",
    "torchsde>=0.2.6",
    "aiorun>=2023.7.2",
]

readme = "README.md"
requires-python = ">= 3.10"

[tool.rye]
managed = true
dev-dependencies = []

[project.scripts]
discord-virtual-micro = "discord_virtual_micro:main"

[[tool.rye.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"

[tool.rye.scripts]
start = { call = "discord_virtual_micro:main" }


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/discord_virtual_micro"]
